+++ src/commands/analyzer.rs
@@ -1,4 +1,4 @@
-//! Implementation of the analyzer command.
+//! Implementation of the language server protocol (LSP) subcommand.
 
 use clap::Parser;
 use wdl::lsp::Server;
@@ -7,7 +7,7 @@ use wdl::lsp::ServerOptions;
 /// Arguments for the `analyzer` subcommand.
 #[derive(Parser, Debug)]
 #[command(author, version, about)]
-pub struct AnalyzerArgs {
+pub struct Args {
     /// Use stdin and stdout for the RPC transport.
     #[clap(long, required = true)]
     pub stdio: bool,
@@ -18,7 +18,7 @@ pub struct AnalyzerArgs {
 }
 
 /// Runs the `analyzer` command.
-pub async fn analyzer(args: AnalyzerArgs) -> anyhow::Result<()> {
+pub async fn analyzer(args: Args) -> anyhow::Result<()> {
     Server::run(ServerOptions {
         name: Some("Sprocket".into()),
         version: Some(env!("CARGO_PKG_VERSION").into()),