+++ src/commands/explain.rs
@@ -1,11 +1,13 @@
-//! Implementation of the explain command.
+//! Implementation of the `explain` subcommand.
 
 use std::collections::HashSet;
+use std::sync::LazyLock;
 
 use anyhow::Ok;
 use anyhow::anyhow;
 use anyhow::bail;
 use clap::Parser;
+use clap::builder::PossibleValuesParser;
 use colored::Colorize;
 use wdl::analysis;
 use wdl::lint;
@@ -16,21 +18,64 @@ const USAGE: &str = "sprocket explain [RULE]
     sprocket explain --tag <TAG>
     sprocket explain --definitions";
 
+/// All rule IDs sorted alphabetically.
+pub static ALL_RULE_IDS: LazyLock<Vec<String>> = LazyLock::new(|| {
+    let mut ids: Vec<String> = analysis::rules()
+        .iter()
+        .map(|r| r.id().to_string())
+        .collect();
+    ids.extend(lint::rules().iter().map(|r| r.id().to_string()));
+    ids.sort();
+    ids
+});
+
+/// All tag names sorted alphabetically.
+static ALL_TAG_NAMES: LazyLock<Vec<String>> = LazyLock::new(|| {
+    let mut tags: HashSet<Tag> = HashSet::new();
+    for rule in lint::rules() {
+        for tag in rule.tags().iter() {
+            tags.insert(tag);
+        }
+    }
+    let mut tag_names: Vec<String> = tags.into_iter().map(|t| t.to_string()).collect();
+    tag_names.sort();
+    tag_names
+});
+
 /// Arguments for the `explain` subcommand.
 #[derive(Parser, Debug)]
 #[command(author, version, about, after_help = generate_after_help(), override_usage = USAGE)]
 pub struct Args {
     /// The name of the rule to explain.
-    #[arg(required_unless_present_any = ["tag", "definitions"], value_name = "RULE")]
+    #[arg(required_unless_present_any = [
+        "tag",
+        "definitions",
+        "list_all_rules",
+        "list_all_tags"
+    ],
+        value_name = "RULE",
+        value_parser = PossibleValuesParser::new(ALL_RULE_IDS.iter())
+    )]
     pub rule_name: Option<String>,
 
     /// List all rules with the given tag.
-    #[arg(short, long, value_name = "TAG", conflicts_with_all = ["rule_name", "definitions"])]
+    #[arg(short, long, value_name = "TAG",
+        conflicts_with_all = ["rule_name", "definitions"],
+        value_parser = PossibleValuesParser::new(ALL_TAG_NAMES.iter())
+    )]
     pub tag: Option<String>,
 
     /// Display general WDL definitions.
     #[arg(long, conflicts_with_all = ["rule_name", "tag"])]
     pub definitions: bool,
+
+    /// Lists all rules and exits.
+    #[arg(long, conflicts_with_all = ["list_all_tags"])]
+    pub list_all_rules: bool,
+
+    /// Lists all tags and exits.
+    #[arg(long, conflicts_with_all = ["list_all_rules"])]
+    pub list_all_tags: bool,
 }
 
 /// Display all rules and tags.
@@ -40,48 +85,21 @@ fn generate_after_help() -> String {
 
 /// Lists all rules as a string for displaying.
 pub fn list_all_rules() -> String {
-    let mut result = "Available rules:".to_owned();
-    let analysis_rules = analysis::rules();
-    let lint_rules = lint::rules();
+    let mut result = String::from("Available rules:");
 
-    let mut indexes = (0..(analysis_rules.len() + lint_rules.len())).collect::<Vec<_>>();
-
-    let id = |index: usize| {
-        if index >= analysis_rules.len() {
-            lint_rules[index - analysis_rules.len()].id()
-        } else {
-            analysis_rules[index].id()
-        }
-    };
-
-    indexes.sort_by(|a, b| id(*a).cmp(id(*b)));
-
-    for index in indexes {
-        result.push_str(&format!("\n  - {}", id(index)));
+    for id in ALL_RULE_IDS.iter() {
+        result.push_str(&format!("\n  - {}", id));
     }
-
     result
 }
 
 /// Lists all tags as a string for displaying.
 pub fn list_all_tags() -> String {
-    let mut result = "Available tags:".to_owned();
-    let lint_rules = lint::rules();
+    let mut result = String::from("Available tags:");
 
-    let mut tags: HashSet<Tag> = HashSet::new();
-    for rule in lint_rules {
-        for tag in rule.tags().iter() {
-            tags.insert(tag);
-        }
-    }
-
-    let mut tags: Vec<Tag> = tags.into_iter().collect();
-    tags.sort_by_key(|tag| tag.to_string());
-
-    for tag in tags {
+    for tag in ALL_TAG_NAMES.iter() {
         result.push_str(&format!("\n  - {}", tag));
     }
-
     result
 }
 
@@ -119,6 +137,16 @@ pub fn pretty_print_analysis_rule(rule: &dyn analysis::Rule) {
 
 /// Explains a lint rule.
 pub fn explain(args: Args) -> anyhow::Result<()> {
+    if args.list_all_rules {
+        println!("{}", list_all_rules());
+        return Ok(());
+    }
+
+    if args.list_all_tags {
+        println!("{}", list_all_tags());
+        return Ok(());
+    }
+
     if args.definitions {
         println!("{}", lint::DEFINITIONS_TEXT);
         return Ok(());
@@ -127,7 +155,7 @@ pub fn explain(args: Args) -> anyhow::Result<()> {
     if let Some(tag) = args.tag {
         let target = tag.parse::<Tag>().map_err(|_| {
             println!("{}\n", list_all_tags());
-            anyhow!("Invalid tag '{}'", tag)
+            anyhow!("invalid tag `{}`", tag)
         })?;
 
         let rules = lint::rules()
@@ -137,7 +165,7 @@ pub fn explain(args: Args) -> anyhow::Result<()> {
 
         if rules.is_empty() {
             println!("{}\n", list_all_tags());
-            bail!("No rules found with the tag `{}`", tag);
+            bail!("no rules found with the tag `{}`", tag);
         } else {
             println!("Rules with the tag `{}`:", tag);
             let mut rule_ids = rules.iter().map(|rule| rule.id()).collect::<Vec<_>>();
@@ -169,14 +197,14 @@ pub fn explain(args: Args) -> anyhow::Result<()> {
                     }
                     None => {
                         println!("{rules}\n", rules = list_all_rules());
-                        bail!("No rule found with the name `{rule_name}`");
+                        bail!("no rule found with the name `{rule_name}`");
                     }
                 }
             }
         }
 
-        Ok(())
-    } else {
-        bail!("Invalid arguments: either a rule_name, a --tag, or --definitions must be provided");
+        return Ok(());
     }
+
+    unreachable!();
 }